Flytter tester for fradrag til egen fil.
