+++ service/src/test/kotlin/no/nav/su/se/bakover/service/revurdering/RevurderingServiceImplTest.kt
@@ -1639,216 +1639,4 @@ internal class RevurderingServiceImplTest {
             )
         }
     }
-
-    @Test
-    fun `lagreFradrag happy case`() {
-        val revurderingId = UUID.randomUUID()
-
-        val revurderingsperiode = Periode.create(1.januar(2021), 31.desember(2021))
-        val revuderingMock = mock<OpprettetRevurdering> {
-            on { id } doReturn revurderingId
-            on { periode } doReturn revurderingsperiode
-            on { grunnlagsdata } doReturn Grunnlagsdata(
-                bosituasjon = listOf(
-                    Grunnlag.Bosituasjon.Fullstendig.Enslig(
-                        id = UUID.randomUUID(),
-                        opprettet = fixedTidspunkt,
-                        periode = revurderingsperiode,
-                        begrunnelse = null,
-                    ),
-                ),
-            )
-        }
-
-        val revurderingRepoMock = mock<RevurderingRepo> {
-            on { hent(any()) } doReturn revuderingMock
-        }
-
-        val grunnlagServiceMock = mock<GrunnlagService> {
-            on { lagreFradragsgrunnlag(any(), any()) }.doNothing()
-        }
-
-        val revurderingService = createRevurderingService(
-            revurderingRepo = revurderingRepoMock,
-            grunnlagService = grunnlagServiceMock,
-        )
-
-        val request = LeggTilFradragsgrunnlagRequest(
-            behandlingId = revurderingId,
-            fradragsrunnlag = listOf(
-                Grunnlag.Fradragsgrunnlag(
-                    fradrag = FradragFactory.ny(
-                        type = Fradragstype.Arbeidsinntekt,
-                        månedsbeløp = 0.0,
-                        periode = revurderingsperiode,
-                        utenlandskInntekt = null,
-                        tilhører = FradragTilhører.BRUKER,
-                    ),
-                ),
-            ),
-        )
-
-        revurderingService.leggTilFradragsgrunnlag(
-            request,
-        ).getOrHandle { throw Exception("k") }
-
-        inOrder(
-            revurderingRepoMock,
-            grunnlagServiceMock,
-        ) {
-            verify(revurderingRepoMock).hent(argThat { it shouldBe revurderingId })
-            verify(grunnlagServiceMock).lagreFradragsgrunnlag(
-                argThat { it shouldBe revurderingId },
-                argThat { it shouldBe request.fradragsrunnlag },
-            )
-            verify(revurderingRepoMock).hent(argThat { it shouldBe revurderingId })
-        }
-
-        verifyNoMoreInteractions(revurderingRepoMock, grunnlagServiceMock)
-    }
-
-    @Test
-    fun `lagreFradrag finner ikke revurdering`() {
-        val revurderingRepoMock = mock<RevurderingRepo> {
-            on { hent(any()) } doReturn null
-        }
-
-        val grunnlagServiceMock = mock<GrunnlagService>()
-
-        val revurderingService = createRevurderingService(
-            revurderingRepo = revurderingRepoMock,
-            grunnlagService = grunnlagServiceMock,
-        )
-
-        val request = LeggTilFradragsgrunnlagRequest(
-            behandlingId = revurderingId,
-            fradragsrunnlag = listOf(
-                Grunnlag.Fradragsgrunnlag(
-                    fradrag = FradragFactory.ny(
-                        type = Fradragstype.Arbeidsinntekt,
-                        månedsbeløp = 0.0,
-                        periode = Periode.create(fraOgMed = 1.januar(2021), tilOgMed = 31.desember(2021)),
-                        utenlandskInntekt = null,
-                        tilhører = FradragTilhører.BRUKER,
-                    ),
-                ),
-            ),
-        )
-
-        revurderingService.leggTilFradragsgrunnlag(
-            request,
-        ).shouldBeLeft(KunneIkkeLeggeTilFradragsgrunnlag.FantIkkeBehandling)
-
-        inOrder(
-            revurderingRepoMock,
-            grunnlagServiceMock,
-        ) {
-            verify(revurderingRepoMock).hent(argThat { it shouldBe revurderingId })
-        }
-
-        verifyNoMoreInteractions(revurderingRepoMock, grunnlagServiceMock)
-    }
-
-    @Test
-    fun `lagreFradrag har en status som gjør at man ikke kan legge til fradrag`() {
-        val revurderingId = UUID.randomUUID()
-
-        val revuderingMock = mock<RevurderingTilAttestering> {
-            on { id } doReturn revurderingId
-        }
-
-        val revurderingRepoMock = mock<RevurderingRepo> {
-            on { hent(any()) } doReturn revuderingMock
-        }
-
-        val grunnlagServiceMock = mock<GrunnlagService>()
-
-        val revurderingService = createRevurderingService(
-            revurderingRepo = revurderingRepoMock,
-            grunnlagService = grunnlagServiceMock,
-        )
-
-        val request = LeggTilFradragsgrunnlagRequest(
-            behandlingId = revurderingId,
-            fradragsrunnlag = listOf(
-                Grunnlag.Fradragsgrunnlag(
-                    fradrag = FradragFactory.ny(
-                        type = Fradragstype.Arbeidsinntekt,
-                        månedsbeløp = 0.0,
-                        periode = Periode.create(fraOgMed = 1.januar(2021), tilOgMed = 31.desember(2021)),
-                        utenlandskInntekt = null,
-                        tilhører = FradragTilhører.BRUKER,
-                    ),
-                ),
-            ),
-        )
-
-        revurderingService.leggTilFradragsgrunnlag(
-            request,
-        ).shouldBeLeft(KunneIkkeLeggeTilFradragsgrunnlag.UgyldigStatus)
-
-        inOrder(
-            revurderingRepoMock,
-            grunnlagServiceMock,
-        ) {
-            verify(revurderingRepoMock).hent(argThat { it shouldBe revurderingId })
-        }
-
-        verifyNoMoreInteractions(revurderingRepoMock, grunnlagServiceMock)
-    }
-
-    @Test
-    fun `validerer fradragsgrunnlag`() {
-        val revurderingsperiode = Periode.create(1.mai(2021), 31.desember(2021))
-        val revurderingMock = mock<OpprettetRevurdering>() {
-
-            on { periode } doReturn revurderingsperiode
-            on { grunnlagsdata } doReturn Grunnlagsdata(
-                bosituasjon = listOf(
-                    Grunnlag.Bosituasjon.Fullstendig.Enslig(
-                        id = UUID.randomUUID(),
-                        opprettet = fixedTidspunkt,
-                        periode = revurderingsperiode,
-                        begrunnelse = null,
-                    ),
-                ),
-            )
-        }
-
-        val revurderingRepoMock = mock<RevurderingRepo> {
-            on { hent(any()) } doReturn revurderingMock
-        }
-
-        val revurderingService = createRevurderingService(
-            revurderingRepo = revurderingRepoMock,
-        )
-
-        val request = LeggTilFradragsgrunnlagRequest(
-            behandlingId = revurderingId,
-            fradragsrunnlag = listOf(
-                Grunnlag.Fradragsgrunnlag(
-                    fradrag = FradragFactory.ny(
-                        type = Fradragstype.Arbeidsinntekt,
-                        månedsbeløp = 0.0,
-                        periode = revurderingsperiode,
-                        utenlandskInntekt = null,
-                        tilhører = FradragTilhører.BRUKER,
-                    ),
-                ),
-                Grunnlag.Fradragsgrunnlag(
-                    fradrag = FradragFactory.ny(
-                        type = Fradragstype.Kontantstøtte,
-                        månedsbeløp = 0.0,
-                        periode = Periode.create(fraOgMed = 1.januar(2021), tilOgMed = 31.juli(2021)),
-                        utenlandskInntekt = null,
-                        tilhører = FradragTilhører.BRUKER,
-                    ),
-                ),
-            ),
-        )
-
-        revurderingService.leggTilFradragsgrunnlag(
-            request,
-        ).shouldBeLeft(KunneIkkeLeggeTilFradragsgrunnlag.FradragsgrunnlagUtenforRevurderingsperiode)
-    }
 }
