+++ service/src/test/kotlin/no/nav/su/se/bakover/service/revurdering/RevurderingServiceLeggTilFradragsgrunnlagTest.kt
@@ -1,0 +1,243 @@
+package no.nav.su.se.bakover.service.revurdering
+
+import arrow.core.getOrHandle
+import com.nhaarman.mockitokotlin2.any
+import com.nhaarman.mockitokotlin2.doReturn
+import com.nhaarman.mockitokotlin2.inOrder
+import com.nhaarman.mockitokotlin2.mock
+import com.nhaarman.mockitokotlin2.verifyNoMoreInteractions
+import io.kotest.assertions.arrow.either.shouldBeLeft
+import io.kotest.matchers.shouldBe
+import no.nav.su.se.bakover.common.desember
+import no.nav.su.se.bakover.common.januar
+import no.nav.su.se.bakover.common.juli
+import no.nav.su.se.bakover.common.mai
+import no.nav.su.se.bakover.common.periode.Periode
+import no.nav.su.se.bakover.database.revurdering.RevurderingRepo
+import no.nav.su.se.bakover.domain.beregning.fradrag.FradragFactory
+import no.nav.su.se.bakover.domain.beregning.fradrag.FradragTilhører
+import no.nav.su.se.bakover.domain.beregning.fradrag.Fradragstype
+import no.nav.su.se.bakover.domain.grunnlag.Grunnlag
+import no.nav.su.se.bakover.domain.grunnlag.Grunnlagsdata
+import no.nav.su.se.bakover.domain.revurdering.OpprettetRevurdering
+import no.nav.su.se.bakover.domain.revurdering.RevurderingTilAttestering
+import no.nav.su.se.bakover.service.argThat
+import no.nav.su.se.bakover.service.doNothing
+import no.nav.su.se.bakover.service.fixedTidspunkt
+import no.nav.su.se.bakover.service.grunnlag.GrunnlagService
+import org.junit.jupiter.api.Test
+import java.util.UUID
+
+class RevurderingServiceLeggTilFradragsgrunnlagTest {
+    @Test
+    fun `lagreFradrag happy case`() {
+        val revurderingId = UUID.randomUUID()
+
+        val revurderingsperiode = Periode.create(1.januar(2021), 31.desember(2021))
+        val revuderingMock = mock<OpprettetRevurdering> {
+            on { id } doReturn revurderingId
+            on { periode } doReturn revurderingsperiode
+            on { grunnlagsdata } doReturn Grunnlagsdata(
+                bosituasjon = listOf(
+                    Grunnlag.Bosituasjon.Fullstendig.Enslig(
+                        id = UUID.randomUUID(),
+                        opprettet = fixedTidspunkt,
+                        periode = revurderingsperiode,
+                        begrunnelse = null,
+                    ),
+                ),
+            )
+        }
+
+        val revurderingRepoMock = mock<RevurderingRepo> {
+            on { hent(any()) } doReturn revuderingMock
+        }
+
+        val grunnlagServiceMock = mock<GrunnlagService> {
+            on { lagreFradragsgrunnlag(any(), any()) }.doNothing()
+        }
+
+        val revurderingService = RevurderingTestUtils.createRevurderingService(
+            revurderingRepo = revurderingRepoMock,
+            grunnlagService = grunnlagServiceMock,
+        )
+
+        val request = LeggTilFradragsgrunnlagRequest(
+            behandlingId = revurderingId,
+            fradragsrunnlag = listOf(
+                Grunnlag.Fradragsgrunnlag(
+                    fradrag = FradragFactory.ny(
+                        type = Fradragstype.Arbeidsinntekt,
+                        månedsbeløp = 0.0,
+                        periode = revurderingsperiode,
+                        utenlandskInntekt = null,
+                        tilhører = FradragTilhører.BRUKER,
+                    ),
+                ),
+            ),
+        )
+
+        revurderingService.leggTilFradragsgrunnlag(
+            request,
+        ).getOrHandle { throw Exception("k") }
+
+        inOrder(
+            revurderingRepoMock,
+            grunnlagServiceMock,
+        ) {
+            verify(revurderingRepoMock).hent(argThat { it shouldBe revurderingId })
+            verify(grunnlagServiceMock).lagreFradragsgrunnlag(
+                argThat { it shouldBe revurderingId },
+                argThat { it shouldBe request.fradragsrunnlag },
+            )
+            verify(revurderingRepoMock).hent(argThat { it shouldBe revurderingId })
+        }
+
+        verifyNoMoreInteractions(revurderingRepoMock, grunnlagServiceMock)
+    }
+
+    @Test
+    fun `lagreFradrag finner ikke revurdering`() {
+        val revurderingRepoMock = mock<RevurderingRepo> {
+            on { hent(any()) } doReturn null
+        }
+
+        val grunnlagServiceMock = mock<GrunnlagService>()
+
+        val revurderingService = RevurderingTestUtils.createRevurderingService(
+            revurderingRepo = revurderingRepoMock,
+            grunnlagService = grunnlagServiceMock,
+        )
+
+        val request = LeggTilFradragsgrunnlagRequest(
+            behandlingId = RevurderingTestUtils.revurderingId,
+            fradragsrunnlag = listOf(
+                Grunnlag.Fradragsgrunnlag(
+                    fradrag = FradragFactory.ny(
+                        type = Fradragstype.Arbeidsinntekt,
+                        månedsbeløp = 0.0,
+                        periode = Periode.create(fraOgMed = 1.januar(2021), tilOgMed = 31.desember(2021)),
+                        utenlandskInntekt = null,
+                        tilhører = FradragTilhører.BRUKER,
+                    ),
+                ),
+            ),
+        )
+
+        revurderingService.leggTilFradragsgrunnlag(
+            request,
+        ).shouldBeLeft(KunneIkkeLeggeTilFradragsgrunnlag.FantIkkeBehandling)
+
+        inOrder(
+            revurderingRepoMock,
+            grunnlagServiceMock,
+        ) {
+            verify(revurderingRepoMock).hent(argThat { it shouldBe RevurderingTestUtils.revurderingId })
+        }
+
+        verifyNoMoreInteractions(revurderingRepoMock, grunnlagServiceMock)
+    }
+
+    @Test
+    fun `lagreFradrag har en status som gjør at man ikke kan legge til fradrag`() {
+        val revurderingId = UUID.randomUUID()
+
+        val revuderingMock = mock<RevurderingTilAttestering> {
+            on { id } doReturn revurderingId
+        }
+
+        val revurderingRepoMock = mock<RevurderingRepo> {
+            on { hent(any()) } doReturn revuderingMock
+        }
+
+        val grunnlagServiceMock = mock<GrunnlagService>()
+
+        val revurderingService = RevurderingTestUtils.createRevurderingService(
+            revurderingRepo = revurderingRepoMock,
+            grunnlagService = grunnlagServiceMock,
+        )
+
+        val request = LeggTilFradragsgrunnlagRequest(
+            behandlingId = revurderingId,
+            fradragsrunnlag = listOf(
+                Grunnlag.Fradragsgrunnlag(
+                    fradrag = FradragFactory.ny(
+                        type = Fradragstype.Arbeidsinntekt,
+                        månedsbeløp = 0.0,
+                        periode = Periode.create(fraOgMed = 1.januar(2021), tilOgMed = 31.desember(2021)),
+                        utenlandskInntekt = null,
+                        tilhører = FradragTilhører.BRUKER,
+                    ),
+                ),
+            ),
+        )
+
+        revurderingService.leggTilFradragsgrunnlag(
+            request,
+        ).shouldBeLeft(KunneIkkeLeggeTilFradragsgrunnlag.UgyldigStatus)
+
+        inOrder(
+            revurderingRepoMock,
+            grunnlagServiceMock,
+        ) {
+            verify(revurderingRepoMock).hent(argThat { it shouldBe revurderingId })
+        }
+
+        verifyNoMoreInteractions(revurderingRepoMock, grunnlagServiceMock)
+    }
+
+    @Test
+    fun `validerer fradragsgrunnlag`() {
+        val revurderingsperiode = Periode.create(1.mai(2021), 31.desember(2021))
+        val revurderingMock = mock<OpprettetRevurdering>() {
+
+            on { periode } doReturn revurderingsperiode
+            on { grunnlagsdata } doReturn Grunnlagsdata(
+                bosituasjon = listOf(
+                    Grunnlag.Bosituasjon.Fullstendig.Enslig(
+                        id = UUID.randomUUID(),
+                        opprettet = fixedTidspunkt,
+                        periode = revurderingsperiode,
+                        begrunnelse = null,
+                    ),
+                ),
+            )
+        }
+
+        val revurderingRepoMock = mock<RevurderingRepo> {
+            on { hent(any()) } doReturn revurderingMock
+        }
+
+        val revurderingService = RevurderingTestUtils.createRevurderingService(
+            revurderingRepo = revurderingRepoMock,
+        )
+
+        val request = LeggTilFradragsgrunnlagRequest(
+            behandlingId = RevurderingTestUtils.revurderingId,
+            fradragsrunnlag = listOf(
+                Grunnlag.Fradragsgrunnlag(
+                    fradrag = FradragFactory.ny(
+                        type = Fradragstype.Arbeidsinntekt,
+                        månedsbeløp = 0.0,
+                        periode = revurderingsperiode,
+                        utenlandskInntekt = null,
+                        tilhører = FradragTilhører.BRUKER,
+                    ),
+                ),
+                Grunnlag.Fradragsgrunnlag(
+                    fradrag = FradragFactory.ny(
+                        type = Fradragstype.Kontantstøtte,
+                        månedsbeløp = 0.0,
+                        periode = Periode.create(fraOgMed = 1.januar(2021), tilOgMed = 31.juli(2021)),
+                        utenlandskInntekt = null,
+                        tilhører = FradragTilhører.BRUKER,
+                    ),
+                ),
+            ),
+        )
+
+        revurderingService.leggTilFradragsgrunnlag(
+            request,
+        ).shouldBeLeft(KunneIkkeLeggeTilFradragsgrunnlag.FradragsgrunnlagUtenforRevurderingsperiode)
+    }
+}
